+++ crates/weaver_semconv/src/v2/event.rs
@@ -7,7 +7,11 @@ use serde::{Deserialize, Serialize};
 
 use crate::{
     group::GroupSpec,
-    v2::{attribute::AttributeRef, signal_id::SignalId, CommonFields},
+    v2::{
+        attribute::{split_attributes_and_groups, AttributeOrGroupRef},
+        signal_id::SignalId,
+        CommonFields,
+    },
 };
 
 /// Defines a new event.
@@ -19,7 +23,7 @@ pub struct Event {
     /// List of attributes that belong to the semantic convention.
     #[serde(default)]
     #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub attributes: Vec<AttributeRef>,
+    pub attributes: Vec<AttributeOrGroupRef>,
     /// Which resources this event should be associated with.
     #[serde(default)]
     #[serde(skip_serializing_if = "Vec::is_empty")]
@@ -33,6 +37,7 @@ impl Event {
     /// Converts a v2 event into a v1 GroupSpec.
     #[must_use]
     pub fn into_v1_group(self) -> GroupSpec {
+        let (attribute_refs, include_groups) = split_attributes_and_groups(self.attributes);
         GroupSpec {
             id: format!("event.{}", &self.name),
             r#type: crate::group::GroupType::Event,
@@ -40,13 +45,10 @@ impl Event {
             note: self.common.note,
             prefix: Default::default(),
             extends: None,
+            include_groups,
             stability: Some(self.common.stability),
             deprecated: self.common.deprecated,
-            attributes: self
-                .attributes
-                .into_iter()
-                .map(|a| a.into_v1_attribute())
-                .collect(),
+            attributes: attribute_refs,
             span_kind: None,
             events: Default::default(),
             metric_name: None,
@@ -61,6 +63,7 @@ impl Event {
                 Some(self.common.annotations)
             },
             entity_associations: self.entity_associations,
+            visibility: None,
         }
     }
 }